+++ CHANGELOG.md
@@ -2,5 +2,15 @@
 
 The project follows Semantic Versioning (http://semver.org/)
 
+## 1.1.0 - 2018-03-12
+### Added
+* HasReferenceInterface::getFilterValues, ReferenceHelper::getClassFilterData
+  and ReferenceHelper::getEntityFilterData to support querying
+  the database for entities that reference objects of a specific class or a
+  concrete objet
+
+### Changed
+* README
+
 ## 1.0.0 - 2018-03-10
 Initial release